+++ lib/yokogaki.dart
@@ -8,6 +8,7 @@ export 'src/widgets/horizontal_text.dart';
 // Models
 export 'src/models/horizontal_text_style.dart';
 export 'src/models/ruby_text.dart';
+export 'src/models/kenten.dart';
 
 // Re-export kinsoku package for convenience
 export 'package:kinsoku/kinsoku.dart';